feat(delete): when deleting post also delete the comment and reply in database

+++ src/server/api/routers/like.ts
@@ -5,23 +5,29 @@ export const likeRouter = createTRPCRouter({
   likePost: protectedProcedure
     .input(
       z.object({
-        id: z.string().optional(),
         postId: z.string(),
         isLiked: z.boolean(),
       })
     )
     .mutation(async ({ ctx, input }) => {
-      if (input.isLiked) {
+      const isLikeExist = await ctx.prisma.likes.findMany({
+        where: {
+          postId: input.postId,
+          userId: ctx.session.user.id,
+        },
+      });
+
+      if (!input.isLiked && isLikeExist.length > 0) {
+        await ctx.prisma.likes.delete({
+          where: {
+            id: isLikeExist[0]?.id,
+          },
+        });
+      } else {
         await ctx.prisma.likes.create({
           data: {
             postId: input.postId,
             userId: ctx.session.user.id,
-          },
-        });
-      } else {
-        await ctx.prisma.likes.delete({
-          where: {
-            id: input.id,
           },
         });
       }
@@ -61,23 +67,29 @@ export const likeRouter = createTRPCRouter({
   likeReplyComment: protectedProcedure
     .input(
       z.object({
-        id: z.string().optional(),
         replyId: z.string(),
         isLiked: z.boolean(),
       })
     )
     .mutation(async ({ ctx, input }) => {
-      if (input.isLiked) {
+      const isLikeExist = await ctx.prisma.likeReplyComment.findMany({
+        where: {
+          replyId: input.replyId,
+          userId: ctx.session.user.id,
+        },
+      });
+
+      if (!input.isLiked && isLikeExist.length > 0) {
+        await ctx.prisma.likeReplyComment.delete({
+          where: {
+            id: isLikeExist[0]?.id,
+          },
+        });
+      } else {
         await ctx.prisma.likeReplyComment.create({
           data: {
             replyId: input.replyId,
             userId: ctx.session.user.id,
-          },
-        });
-      } else {
-        await ctx.prisma.likeReplyComment.delete({
-          where: {
-            id: input.id,
           },
         });
       }
